load-tests/chat/requirements.txt
locust
pika
aio-pika
anyio
python-dotenv